requirements.txt
wandb 
codetiming 
cython==0.29.36 
debugpy 
dill 
draccus 
egl_probe 
GPUtil 
h5py 
hydra-core 
imageio[ffmpeg] 
json_numpy 
jsonlines 
jwt 
matplotlib 
mujoco 
numba 
numpy==1.26.4 
ninja 
nvitop 
omegaconf 
packaging 
pandas 
peft==0.11.0 
protobuf 
psutil 
PyOpenGL==3.1.9 
PyOpenGL_accelerate==3.1.9 
pybind11 
pylatexenc 
pysnooper 
ray[default] 
scipy 
scikit-learn 
tensorflow==2.19.0 
tensorflow_datasets 
tensordict 
timm==0.9.10 
tokenizers==0.19.1 
torchdata 
opencv-python 
webdataset 
xformers==v0.0.28.post3 
diffusers==0.33.0
flash-attn==2.5.5